+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## HEAD (Unreleased)
 
+- feat: Download CLI if preinstalled version has a known issue.
+  ([#956](https://github.com/pulumi/actions/pull/956))
+
 --
 
 ## 4.3.0 (2023-05-12)